+++ com-mmxlabs-models-lng/com.mmxlabs.models.lng.fleet/src/com/mmxlabs/models/lng/fleet/VesselStateAttributes.java
@@ -38,13 +38,13 @@ public interface VesselStateAttributes extends MMXObject {
 	 * </p>
 	 * <!-- end-user-doc -->
 	 * @return the value of the '<em>Nbo Rate</em>' attribute.
-	 * @see #setNboRate(int)
+	 * @see #setNboRate(double)
 	 * @see com.mmxlabs.models.lng.fleet.FleetPackage#getVesselStateAttributes_NboRate()
 	 * @model required="true"
 	 *        annotation="http://www.mmxlabs.com/models/ui/numberFormat unit='m3/d'"
 	 * @generated
 	 */
-	int getNboRate();
+	double getNboRate();
 
 	/**
 	 * Sets the value of the '{@link com.mmxlabs.models.lng.fleet.VesselStateAttributes#getNboRate <em>Nbo Rate</em>}' attribute.
@@ -54,7 +54,7 @@ public interface VesselStateAttributes extends MMXObject {
 	 * @see #getNboRate()
 	 * @generated
 	 */
-	void setNboRate(int value);
+	void setNboRate(double value);
 
 	/**
 	 * Returns the value of the '<em><b>Idle NBO Rate</b></em>' attribute.
@@ -65,13 +65,13 @@ public interface VesselStateAttributes extends MMXObject {
 	 * </p>
 	 * <!-- end-user-doc -->
 	 * @return the value of the '<em>Idle NBO Rate</em>' attribute.
-	 * @see #setIdleNBORate(int)
+	 * @see #setIdleNBORate(double)
 	 * @see com.mmxlabs.models.lng.fleet.FleetPackage#getVesselStateAttributes_IdleNBORate()
 	 * @model required="true"
 	 *        annotation="http://www.mmxlabs.com/models/ui/numberFormat unit='m3/d'"
 	 * @generated
 	 */
-	int getIdleNBORate();
+	double getIdleNBORate();
 
 	/**
 	 * Sets the value of the '{@link com.mmxlabs.models.lng.fleet.VesselStateAttributes#getIdleNBORate <em>Idle NBO Rate</em>}' attribute.
@@ -81,7 +81,7 @@ public interface VesselStateAttributes extends MMXObject {
 	 * @see #getIdleNBORate()
 	 * @generated
 	 */
-	void setIdleNBORate(int value);
+	void setIdleNBORate(double value);
 
 	/**
 	 * Returns the value of the '<em><b>Idle Base Rate</b></em>' attribute.
@@ -92,13 +92,13 @@ public interface VesselStateAttributes extends MMXObject {
 	 * </p>
 	 * <!-- end-user-doc -->
 	 * @return the value of the '<em>Idle Base Rate</em>' attribute.
-	 * @see #setIdleBaseRate(int)
+	 * @see #setIdleBaseRate(double)
 	 * @see com.mmxlabs.models.lng.fleet.FleetPackage#getVesselStateAttributes_IdleBaseRate()
 	 * @model required="true"
 	 *        annotation="http://www.mmxlabs.com/models/ui/numberFormat unit='MT/d'"
 	 * @generated
 	 */
-	int getIdleBaseRate();
+	double getIdleBaseRate();
 
 	/**
 	 * Sets the value of the '{@link com.mmxlabs.models.lng.fleet.VesselStateAttributes#getIdleBaseRate <em>Idle Base Rate</em>}' attribute.
@@ -108,7 +108,7 @@ public interface VesselStateAttributes extends MMXObject {
 	 * @see #getIdleBaseRate()
 	 * @generated
 	 */
-	void setIdleBaseRate(int value);
+	void setIdleBaseRate(double value);
 
 	/**
 	 * Returns the value of the '<em><b>In Port Base Rate</b></em>' attribute.
@@ -119,13 +119,13 @@ public interface VesselStateAttributes extends MMXObject {
 	 * </p>
 	 * <!-- end-user-doc -->
 	 * @return the value of the '<em>In Port Base Rate</em>' attribute.
-	 * @see #setInPortBaseRate(int)
+	 * @see #setInPortBaseRate(double)
 	 * @see com.mmxlabs.models.lng.fleet.FleetPackage#getVesselStateAttributes_InPortBaseRate()
 	 * @model required="true"
 	 *        annotation="http://www.mmxlabs.com/models/ui/numberFormat unit='MT/d'"
 	 * @generated
 	 */
-	int getInPortBaseRate();
+	double getInPortBaseRate();
 
 	/**
 	 * Sets the value of the '{@link com.mmxlabs.models.lng.fleet.VesselStateAttributes#getInPortBaseRate <em>In Port Base Rate</em>}' attribute.
@@ -135,7 +135,7 @@ public interface VesselStateAttributes extends MMXObject {
 	 * @see #getInPortBaseRate()
 	 * @generated
 	 */
-	void setInPortBaseRate(int value);
+	void setInPortBaseRate(double value);
 
 	/**
 	 * Returns the value of the '<em><b>Fuel Consumption</b></em>' containment reference list.
